Fix comparison of int arrays with out-of-bounds pyints (#350)
This adds a special path to allow comparisons with out-of-bounds Python integers. NumPy goes a little bit more these days (it defines int64 to uint64 comparisons correctly, but this is not a NumPy 2 change).

This was the easiest solution I could come up with, but maybe someone deeper in cunumeric has a better idea.